requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==4.2.14
django-filter==24.3
djangorestframework==3.15.2
drf-spectacular==0.27.2
django-debug-toolbar==4.4.6
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
pillow==10.4.0
psycopg==3.2.2
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1